agent_framework/pyproject.toml
[project]
name = "qwen-agent"
version = "0.1.0"
description = "A powerful agent framework for Ollama/Qwen models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.3.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
qwen-agent = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
